sync: 同步 v3-cloud/v3-frame 最新代码
- v3-core: DataMaskProcessor/DataMaskAspect/DateUtils/HttpServletUtil 更新
- v3-data: 移除 postgresql/ojdbc8 驱动
- v3-log: 移除 transmittable-thread-local/jetbrains annotations
- v3-web: 移除 ApiIdempotent/IdempotentAspect 及 hutool-crypto
- v3-secure: AuditLog/权限枚举/LoginUser 更新
- hj-parent: 清理 guava/transmittable/jetbrains/hutool-crypto 管理项

+++ v3-web/src/main/java/com/wenx/handler/MybatisPlusMetaObjectHandler.java
@@ -39,9 +39,9 @@ public class MybatisPlusMetaObjectHandler implements MetaObjectHandler {
                 this.strictInsertFill(metaObject, "version", Integer.class, 1);
             }
             
-            // 删除标识
+            // 删除标识（Z4：Long 型，删除时置为行 id，见 BaseEntity）
             if (metaObject.hasSetter("deleted")) {
-                this.strictInsertFill(metaObject, "deleted", Integer.class, 0);
+                this.strictInsertFill(metaObject, "deleted", Long.class, 0L);
             }
             
             // 创建者和修改者
